superbot_platform/requirements.txt
# Dependências do SuperBot Platform
# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Configuração
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Banco de Dados
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Cache
redis>=5.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Google Gemini
google-genai>=1.0.0

# ElevenLabs
elevenlabs>=1.0.0

# OpenRouter (compatibilidade OpenAI)
openai>=1.6.0

# Twilio (ligações)
twilio>=8.10.0

# Utilitários
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
